refactor!: Reorganize exceptions into domain packages
* Placing exceptions in packages along side related classes
* Change ConnectException to extend PathRequestException
