+++ common/src/main/java/com/mx/common/accessors/AccessorMethodNotImplementedException.java
@@ -1,6 +1,6 @@
-package com.mx.common.exception;
+package com.mx.common.accessors;
 
-import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.PathRequestException;
 
 /**
  * Thrown when an accessor method is invoked that has no implementation.
