+++ common/src/main/java/com/mx/common/accessors/ResourceNotFoundException.java
@@ -1,6 +1,6 @@
-package com.mx.common.exception;
+package com.mx.common.accessors;
 
-import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.PathRequestException;
 
 /**
  * Thrown when a resource is requested that does not exist.
