+++ common/src/main/java/com/mx/common/accessors/ResponsePayloadException.java
@@ -1,4 +1,6 @@
-package com.mx.common.exception;
+package com.mx.common.accessors;
+
+import com.mx.common.exception.PathRequestException;
 
 /**
  * Thrown when response payload is unrecognizable or un-processable
