+++ common/src/main/java/com/mx/common/configuration/InvalidStateException.java
@@ -1,4 +1,6 @@
-package com.mx.common.exception;
+package com.mx.common.configuration;
+
+import com.mx.common.exception.PathSystemException;
 
 /**
  * Thrown when invalid code state is detected, at boot time.
