+++ common/src/main/java/com/mx/common/connect/ServiceUnavailableException.java
@@ -1,6 +1,7 @@
-package com.mx.common.exception;
+package com.mx.common.connect;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.PathRequestException;
 
 /**
  * Thrown when a service (typically upstream service) is unavailable. The unavailability could be determined by a response
