+++ common/src/main/java/com/mx/common/connect/TimeoutException.java
@@ -1,6 +1,7 @@
-package com.mx.common.exception;
+package com.mx.common.connect;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.PathRequestException;
 
 /**
  * Thrown on a connection/request timeout
