+++ common/src/main/java/com/mx/common/facility/FacilityException.java
@@ -1,22 +1,23 @@
-package com.mx.common.exception;
+package com.mx.common.facility;
 
 import com.mx.common.accessors.PathResponseStatus;
+import com.mx.common.exception.PathRequestException;
 
 /**
- * Thrown on unrecoverable error in behavior code
+ * Thrown on unrecoverable error in facility code
  *
  * <p>
  *   See {@link PathRequestException} for usage details
  * </p>
  */
-public class BehaviorException extends GatewayException {
-  public BehaviorException(String message) {
+public class FacilityException extends PathRequestException {
+  public FacilityException(String message) {
     super(message);
     setReport(true);
     setStatus(PathResponseStatus.INTERNAL_ERROR);
   }
 
-  public BehaviorException(String message, Throwable cause) {
+  public FacilityException(String message, Throwable cause) {
     super(message, cause);
     setReport(true);
     setStatus(PathResponseStatus.INTERNAL_ERROR);
