+++ gateway/src/main/java/com/mx/path/gateway/configuration/AccessorProxyMap.java
@@ -7,7 +7,7 @@ import java.util.Locale;
 import java.util.Map;
 
 import com.mx.common.accessors.Accessor;
-import com.mx.common.exception.GatewayException;
+import com.mx.common.gateway.GatewayException;
 import com.mx.path.gateway.configuration.annotations.AccessorScope;
 
 public class AccessorProxyMap {
