+++ gateway/src/main/java/com/mx/path/gateway/configuration/ConnectionConstructionContext.java
@@ -10,7 +10,7 @@ import lombok.Getter;
 
 import com.mx.common.configuration.Configuration;
 import com.mx.common.connect.AccessorConnectionSettings;
-import com.mx.common.exception.GatewayException;
+import com.mx.common.gateway.GatewayException;
 import com.mx.common.reflection.Constructors;
 import com.mx.common.reflection.Fields;
 import com.mx.path.gateway.configuration.annotations.ClientID;
