+++ gateway/src/test/groovy/com/mx/path/gateway/configuration/AccessorProxyMapTest.groovy
@@ -1,6 +1,6 @@
 package com.mx.path.gateway.configuration
 
-import com.mx.common.exception.GatewayException
+import com.mx.common.gateway.GatewayException
 import com.mx.testing.AccountAccessorImpl
 import com.mx.testing.BaseAccessorImpl
 import com.mx.testing.ChildAccessorAccount
