+++ http/src/main/java/com/mx/path/api/connect/http/HttpClientFilter.java
@@ -10,10 +10,10 @@ import java.util.List;
 import com.google.gson.Gson;
 import com.mx.common.collections.MultiValueMap;
 import com.mx.common.collections.SingleValueMap;
+import com.mx.common.connect.ConnectException;
 import com.mx.common.connect.Request;
 import com.mx.common.connect.RequestFilterBase;
 import com.mx.common.connect.Response;
-import com.mx.common.exception.ConnectException;
 import com.mx.common.http.HttpStatus;
 import com.mx.path.api.connect.http.certificates.MutualAuthProvider;
 import com.mx.path.api.connect.http.certificates.MutualAuthProviderFactory;
