+++ test-models/src/main/java/com/mx/testing/accessors/TransactionBaseAccessor.java
@@ -3,8 +3,8 @@ package com.mx.testing.accessors;
 import com.mx.common.accessors.API;
 import com.mx.common.accessors.Accessor;
 import com.mx.common.accessors.AccessorConfiguration;
+import com.mx.common.accessors.AccessorMethodNotImplementedException;
 import com.mx.common.accessors.AccessorResponse;
-import com.mx.common.exception.AccessorMethodNotImplementedException;
 import com.mx.common.gateway.GatewayAPI;
 import com.mx.common.gateway.GatewayClass;
 import com.mx.common.models.MdxList;
